fix again 3

+++ tasks/mpi/kalinin_d_vector_dot_product/src/ops_mpi.cpp
@@ -84,7 +84,7 @@ bool kalinin_d_vector_dot_product_mpi::TestMPITaskParallel::PreProcessingImpl() 
       counts_[i] = delta + (i < remainder ? 1 : 0);  // Assign 1 additional element to the first 'remainder' processes
     }
   }
-  boost::mpi::broadcast(world_, counts_.data(), static_cast<int>(num_processes_), 0);
+  boost::mpi::broadcast(world_, counts_.data(), num_processes_, 0);
 
   if (world_.rank() == 0) {
     input_ = std::vector<std::vector<int>>(task_data->inputs.size());
